feat(tasks): utxo

+++ control/src/configs/defaults.rs
@@ -5,7 +5,7 @@ use big_schema::{App, Storage};
 pub const NUM_KEYS: u64 = 10_000_000;
 pub const READ_RATIO: f64 = 0.5;
 pub const NUM_FAULTY_NODES: u16 = 1;
-pub const NUM_CONCURRENT: u64 = 1;
+pub const NUM_CONCURRENT: u32 = 1;
 
 pub const STORAGE: Storage = Storage::Full;
 pub const APP: App = App::Ycsb;
